% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/associatedTargets.R
\name{associatedTargets}
\alias{associatedTargets}
\title{Get associated targets for a disease category}
\usage{
associatedTargets(efoid, n)
}
\description{
Retrive n drug targets associated with disease with specified EFO id
}
\details{
@param efoid A string. The Experimental Factor Ontology (EFO id)
 @param n  An integer. Number of the record to return
 @return Returns a dataframe which includes data table of the associated drug targets
 @examples
 \dontrun{associatedTargets("EFO_0000540",10)}
 @export
}
