% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colocalisationsForGene.R
\name{colocalisationsForGene}
\alias{colocalisationsForGene}
\title{Get colocalization data for a gene}
\usage{
colocalisationsForGene(ensmbl_ids)
}
\arguments{
\item{ensmbl_ids}{is a identification id for genes by ensembl database}
}
\value{
A dataframe including the queried gene indentity and its colocalization data
}
\description{
Get colocalization data for a gene
}
\examples{
colocalisationsForGene(list("ENSG00000163946","ENSG00000169174", "ENSG00000143001"))
colocalisationsForGene("ENSG00000169174")
}
