% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compGenomicsQuery.R
\name{compGenomicsQuery}
\alias{compGenomicsQuery}
\title{Retrieve Comparative Genomics data for a specified gene.}
\usage{
compGenomicsQuery(ensemblId)
}
\arguments{
\item{ensemblId}{Character: ENSEMBL ID of the target gene (e.g., ENSG00000169174).}
}
\value{
Returns a data frame containing comparative genomics data for the specified gene.
}
\description{
This function queries the Open Targets Genetics GraphQL API to retrieve comparative genomics data
for a specified gene.
}
\examples{
\dontrun{
result <- compGenomicsQuery(ensemblId = "ENSG00000169174")
}
}
