% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneInfo.R
\name{geneInfo}
\alias{geneInfo}
\title{Retrieves gene information for a gene ENSEMBL identifier or a gene name.}
\usage{
geneInfo(gene)
}
\arguments{
\item{gene}{String: a ensembl gene identifier or gene name (e.g.ENSG00000169174/PCSK9).}
}
\value{
Returns a tibble data table containing the following columns:

\enumerate{
\item id
\item symbol
\item bioType
\item description
\item chromosome
\item tss
\item start
\item end
\item fwdStrand
\item exons
}
}
\description{
This function takes gene ensembl id (e.g.ENSG00000169174) or gene name (e.g.PCSK9).
}
\examples{
\dontrun{
otargen::geneInfo(gene="ENSG00000169174")
otargen::geneInfo(gene="PCSK9")
}
}
