% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genes.R
\name{genes}
\alias{genes}
\title{Information about genes on an input locus}
\usage{
genes(chromosome, start, end)
}
\arguments{
\item{chromosome}{Chromosome given as a string}

\item{start}{start position of the input chromosome}

\item{end}{end position of the input chromosome}
}
\value{
A dataframe containing the details of all the genes in the mentioned locus.
}
\description{
Information about genes on an input locus
}
\examples{
genes(chromosome="1", start=800, end=500000)
}
