% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genesForVariant.R
\name{genesForVariant}
\alias{genesForVariant}
\title{Variant to Gene (V2G) data}
\usage{
genesForVariant(variantid)
}
\arguments{
\item{variantid}{is the Open Target Genetics generated id for each variants.}
}
\value{
A dataframe containing variant to gene (v2g) information with individual QTL associations, distances and functional predictions.
}
\description{
Variant to Gene (V2G) data
}
\examples{
genesForVariant("1_55039974_G_T")
}
