% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneticConstraintQuery.R
\name{geneticConstraintQuery}
\alias{geneticConstraintQuery}
\title{Retrieve Genetic Constraint data for a specified gene.}
\usage{
geneticConstraintQuery(ensemblId)
}
\arguments{
\item{ensemblId}{Character: ENSEMBL ID of the target gene (e.g., ENSG00000169174).}
}
\value{
Returns a data frame containing genetic constraint data for the specified gene.
}
\description{
This function queries the Open Targets Genetics GraphQL API to retrieve genetic constraint data
for a specified gene.
}
\examples{
\dontrun{
result <- geneticConstraintQuery(ensemblId = "ENSG00000169174")
}
}
