% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneticEvidenceQuery.R
\name{geneticEvidenceQuery}
\alias{geneticEvidenceQuery}
\title{Query genetic evidance for a gene in a disease}
\usage{
geneticEvidenceQuery(efo, ensid)
}
\arguments{
\item{efo}{is experimental factor ontolody id}

\item{ensid}{is a identification id for genes by ensembl database}
}
\value{
a dataframe including the queried gene indentity and its colocalization data
}
\description{
Query genetic evidance for a gene in a disease
}
