% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_genes.R
\name{get_genes}
\alias{get_genes}
\title{Retrieves information about genes on an input locus}
\usage{
get_genes(chromosome, start, end)
}
\arguments{
\item{\emph{chromosome}}{String: chromosome number as string.}

\item{\emph{start}}{Long: start position of the specified chromosome.}

\item{\emph{end}}{Long: end position of the specified chromosome.}
}
\value{
Returns a tibble data table of all the overlapping genes in the specified region with following information columns:

\enumerate{
\item id
\item symbol
\item bioType
\item description
\item chromosome
\item tss
\item start
\item end
\item fwdStrand
\item exons
}
}
\description{
This function gets information for a region of a specified chromosome .
}
\examples{
\dontrun{
otargen::get_genes(chromosome = "2", start = 239634984, end = 241634984)
}
}
