% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwasColocalisation.R
\name{gwasColocalisation}
\alias{gwasColocalisation}
\title{Get gwas colocalisation for a variant in a study}
\usage{
gwasColocalisation(studyid, variantid)
}
\arguments{
\item{studyid}{is the Open Target Genetics generated id for gwas studies.}

\item{variantid}{is Open Target Genetics generated id for each variant in the database.}
}
\value{
A data frame of the studies which colocalise with the input variant and study.
}
\description{
Get gwas colocalisation for a variant in a study
}
\examples{
gwasColocalisation("GCST90002357", "1_154119580_C_A")
}
