% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwasColocalisationForRegion.R
\name{gwasColocalisationForRegion}
\alias{gwasColocalisationForRegion}
\title{Retrieves GWAS colocalisation data for a region.}
\usage{
gwasColocalisationForRegion(chromosome, start, end)
}
\arguments{
\item{chromosome}{String: Chromosome number as a string.}

\item{start}{Long: Start position of the specified chromosome.}

\item{end}{Long: End position of the specified chromosome.}
}
\value{
Returns a data frame with the following columns:
\itemize{
  \item{\code{study.studyId}:} \emph{Character vector}. Study identifier.
  \item{\code{study.traitReported}:} \emph{Character vector}. Reported trait associated with the colocalisation.
  \item{\code{study.traitCategory}:} \emph{Character vector}. Category of the reported trait.
  \item{\code{indexVariant.id}:} \emph{Character vector}. ID of the index variant.
  \item{\code{indexVariant.position}:} \emph{Integer vector}. Position of the index variant.
  \item{\code{indexVariant.chromosome}:} \emph{Character vector}. Chromosome of the index variant.
  \item{\code{indexVariant.rsId}:} \emph{Character vector}. rsID of the index variant.
  \item{\code{beta}:} \emph{Numeric vector}. Beta value associated with the colocalisation.
  \item{\code{h3}:} \emph{Numeric vector}. H3 value associated with the colocalisation.
  \item{\code{h4}:} \emph{Numeric vector}. H4 value associated with the colocalisation.
  \item{\code{log2h4h3}:} \emph{Numeric vector}. Log2 ratio of H4 to H3 values.
}
}
\description{
Given a defined chromosomal region, this function returns a tibble data table
of colocalised variants to the left and right side of the provided region, including
the calculated colocalization scores.
}
\examples{
\dontrun{
gwasColocalisationForRegion(chromosome = "1", start = 153992685, end = 154155116)
}
}
\references{
- Giambartolomei, Claudia et al. “Bayesian test for colocalisation between pairs of
genetic association studies using summary statistics.” PLoS genetics vol. 10,5 e1004383. 15 May. 2014, doi:10.1371/journal.pgen.1004383
}
