% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwasCredibleSet.R
\name{gwasCredibleSet}
\alias{gwasCredibleSet}
\title{Get GWAS credible set data for a variant in a study}
\usage{
gwasCredibleSet(studyid, variantid)
}
\arguments{
\item{\emph{studyid}}{String: Open Target Genetics generated ID for the GWAS study.}

\item{\emph{variantid}}{String: Open Target Genetics generated ID for the variant (CHRPOSITION_REFALLELE_ALTALLELE or rsId).}
}
\value{
Returns a data frame of results from the credible set of variants for a specific lead variant with the following columns:
\itemize{
  \item{\code{tagVariant.id}:} \emph{Data frame}. A table of IDs of the tag variant.
  \item{\code{tagVariant.rsId}:} \emph{Character vector}. rsID of the tag variant.
  \item{\code{beta}:} \emph{Numeric}. Beta value.
  \item{\code{postProb}:} \emph{Numeric}. Posterior probability.
  \item{\code{pval}:} \emph{Numeric}. P-value.
  \item{\code{se}:} \emph{Numeric}. Standard error.
  \item{\code{MultisignalMethod}:} \emph{Character vector}. Multisignal method.
  \item{\code{logABF}:} \emph{Numeric}. Logarithm of approximate Bayes factor.
  \item{\code{is95}:} \emph{Logical}. Indicates if the variant has a 95% credible set.
  \item{\code{is99}:} \emph{Logical}. Indicates if the variant has a 99% credible set.
}
}
\description{
Provided with a study ID and a lead variant ID, this function returns a
data frame consisting of all the associated credible set tag variants
 with the corresponding statistical data.
}
\examples{
\dontrun{
result <- gwasCredibleSet(studyid="GCST90002357", variantid="1_154119580_C_A")
result <- gwasCredibleSet(studyid="GCST90002357", variantid="rs2494663")
}
}
