% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwasRegional.R
\name{gwasRegional}
\alias{gwasRegional}
\title{Retrieve all signals for a specific region of a chromosome in a GWAS .}
\usage{
gwasRegional(studyid, chromosome, start, end)
}
\arguments{
\item{\emph{studyid}}{String: Open Target Genetics generated ID for the GWAS study.}

\item{\emph{chromosome}}{String: Chromosome number as a string.}

\item{\emph{start}}{Long: Start position of the specified chromosome.}

\item{\emph{end}}{Long: End position of the specified chromosome.}
}
\value{
Returns a data table of variant information and p-values with the following columns:
\itemize{
  \item{\code{variant.id}:} \emph{Character vector}. Variant identifier.
  \item{\code{variant.chromosome}:} \emph{Character vector}. Chromosome of the variant.
  \item{\code{variant.position}:} \emph{Integer vector}. Position of the variant.
  \item{\code{pval}:} \emph{Numeric vector}. P-value.
}
}
\description{
For a given study ID and chromosomal region information, this function returns a tibble data table with all variants and their respective p-values.
}
\examples{
\dontrun{
result <- gwasRegional(studyid = "GCST90002357", chromosome = "1", start = 153992685, end = 154155116)
}
}
