% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hallmarksQuery.R
\name{hallmarksQuery}
\alias{hallmarksQuery}
\title{Retrieve Cancer Hallmarks data for a specified gene.}
\usage{
hallmarksQuery(ensgId)
}
\arguments{
\item{ensgId}{Character: ENSEMBL ID of the target gene (e.g., ENSG00000141510).}
}
\value{
Returns a tibble containing cancer hallmarks data for the specified gene.
}
\description{
This function queries the Open Targets GraphQL API to retrieve cancer hallmarks data
for a specified gene.
}
\examples{
\dontrun{
result <- hallmarksQuery(ensgId = "ENSG00000141510")
}
}
