% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indexVariantsAndStudiesForTagVariant.R
\name{indexVariantsAndStudiesForTagVariant}
\alias{indexVariantsAndStudiesForTagVariant}
\title{Retrieves index variants and studies for a tag variant.}
\usage{
indexVariantsAndStudiesForTagVariant(variantid, pageindex = 0, pagesize = 20)
}
\arguments{
\item{\emph{variantid}}{String: Open Target Genetics generated id for variant (CHR_POSITION_REFALLELE_ALT_ALLELE or rsId).}

\item{\emph{pageindex}}{Int: Index of the current page, pagination index >= 0.}

\item{\emph{pagesize}}{Int: No. of records in a page, pagination size > 0.}
}
\value{
Returns a dataframe containing the variant associated to the input tag variant. The table consists of the following columns:

\enumerate{
\item index_variant
\item study
\item pval
\item pval_mantissa
\item pval_exponent
\item n_total
\item n_cases
\item overall_r2
\item afr1000g_prop
\item amr1000g_prop
\item eas1000g_prop
\item eur1000g_prop
\item sas1000g_prop
\item log10abf
\item posterior_probability
\item odds_ratio
\item odds_ratio_ci_lower
\item odds_ratio_ci_upper
\item beta
\item beta_ci_lower
\item beta_ci_upper
\item direction
}
}
\description{
For an input tag variant id, this function returns a  data table in tibble format
with population level summary stats data columns across various GWAS studies.
}
\examples{
\dontrun{
indexVariantsAndStudiesForTagVariant(variantid = "1_109274968_G_T")
indexVariantsAndStudiesForTagVariant(variantid = "rs12740374",
 pageindex = 1, pagesize = 50)
}
}
