% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicationsQuery.R
\name{indicationsQuery}
\alias{indicationsQuery}
\title{Retrieve Indications data for a specified drug.}
\usage{
indicationsQuery(chemblId)
}
\arguments{
\item{chemblId}{Character: ChEMBL ID of the target drug (e.g., "CHEMBL1016").}
}
\value{
Returns a tibble containing indications data for the specified drug.
}
\description{
This function queries the Open Targets GraphQL API to retrieve indications data
for a specified drug.
}
\examples{
\dontrun{
result <- indicationsQuery(chemblId = "CHEMBL1016")
}
}
