% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knownDrugsGeneQuery.R
\name{knownDrugsGeneQuery}
\alias{knownDrugsGeneQuery}
\title{Retrieve drug and clinical candidate data for a specified gene.}
\usage{
knownDrugsGeneQuery(ensgId)
}
\arguments{
\item{ensgId}{Character: ENSEMBL ID of the target gene (e.g., ENSG00000169174).}
}
\value{
Returns a data frame containing drug and clinical candidate data for the specified gene.
}
\description{
This function queries the Open Targets GraphQL API to retrieve drug
and clinical candidate data for a specified gene.
}
\examples{
\dontrun{
result <- knownDrugsGeneQuery(ensgId = "ENSG00000169174")
}
}
