% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manhattan.R
\name{manhattan}
\alias{manhattan}
\title{Gives list of manhattan association for a given study.}
\usage{
manhattan(studyid, pageindex = 0, pagesize = 0)
}
\arguments{
\item{studyid}{is the Open Target Genetics generated id for gwas studies.}

\item{pageIndex}{optional argument}

\item{pageSize}{optional argument}
}
\description{
Gives list of manhattan association for a given study.
}
