% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mechanismsOfActionQuery.R
\name{mechanismsOfActionQuery}
\alias{mechanismsOfActionQuery}
\title{Retrieve Mechanisms of Action data for a specified drug.}
\usage{
mechanismsOfActionQuery(chemblId)
}
\arguments{
\item{chemblId}{Character: ChEMBL ID of the target drug (e.g., "CHEMBL1016").}
}
\value{
Returns a tibble containing mechanisms of action data for the specified drug.
}
\description{
This function queries the Open Targets GraphQL API to retrieve mechanisms of action data
for a specified drug.
}
\examples{
\dontrun{
result <- mechanismsOfActionQuery(chemblId = "CHEMBL1016")
}
}
