% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mousePhenotypesQuery.R
\name{mousePhenotypesQuery}
\alias{mousePhenotypesQuery}
\title{Retrieve Mouse Phenotypes data for a specified gene.}
\usage{
mousePhenotypesQuery(ensemblId)
}
\arguments{
\item{ensemblId}{Character: ENSEMBL ID of the target gene (e.g., ENSG00000169174).}
}
\value{
Returns a data frame containing mouse phenotypes data for the specified gene.
}
\description{
This function queries the Open Targets Genetics GraphQL API to retrieve mouse phenotypes data
for a specified gene.
}
\examples{
\dontrun{
result <- mousePhenotypesQuery(ensemblId = "ENSG00000169174")
}
}
