% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathwaysQuery.R
\name{pathwaysQuery}
\alias{pathwaysQuery}
\title{Retrieve Pathways data for a specified gene.}
\usage{
pathwaysQuery(ensgId)
}
\arguments{
\item{ensgId}{Character: ENSEMBL ID of the target gene (e.g., ENSG00000105397).}
}
\value{
Returns a tibble containing pathways data for the specified gene.
}
\description{
This function queries the Open Targets GraphQL API to retrieve pathways data
for a specified gene.
}
\examples{
\dontrun{
result <- pathwaysQuery(ensgId = "ENSG00000105397")
}
}
