% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pheWAS.R
\name{pheWAS}
\alias{pheWAS}
\title{Retrieves Phenome Wide Association Studies for a variant.}
\usage{
pheWAS(variantid)
}
\arguments{
\item{\emph{variantid}}{String: Open Targets Genetics generated ID for each variant.}
}
\value{
Returns a data frame with PheWAS associations consisting of the following columns:
\itemize{
  \item{\code{totalGWASStudies}:} \emph{Integer}. Total number of GWAS studies.
  \item{\code{pval}:} \emph{Numeric}. P-value.
  \item{\code{beta}:} \emph{Numeric}. Beta value.
  \item{\code{oddsRatio}:} \emph{Numeric}. Odds ratio.
  \item{\code{nTotal}:} \emph{Integer}. Total number of participants.
  \item{\code{study.studyId}:} \emph{Character vector}. Study ID.
  \item{\code{study.source}:} \emph{Character vector}. Source of the study.
  \item{\code{study.pmid}:} \emph{Character vector}. PMID (PubMed ID) of the study.
  \item{\code{study.pubDate}:} \emph{Character vector}. Publication date of the study.
  \item{\code{study.traitReported}:} \emph{Character vector}. Reported trait.
  \item{\code{study.traitCategory}:} \emph{Character vector}. Trait category.
}
}
\description{
PheWAS (Phenome-wide association study) is a method that investigates the relationships
between genetic variants and traits or phenotypes, helping in the study of their
potential influence on multiple traits or diseases concurrently.
}
\examples{
\dontrun{
result <- pheWAS(variantid = "1_154549918_C_A")
result <- pheWAS(variantid = "rs72698179")
}
}
