% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_coloc.R
\name{plot_coloc}
\alias{plot_coloc}
\title{Plot the colocalisation function results}
\usage{
plot_coloc(data, biobank = FALSE)
}
\arguments{
\item{\emph{data}}{Data Frame: result of colocalisationsForGene function in data
frame format, contacting the phewas information for a variant id}

\item{\emph{biobank}}{Logical: TRUE and FALSE variable, with the default value of
FALSE which will keep the data that are from UKbioBank beside the published GWAS data. In case,
this parameter is set to TRUE, only UKbioBank data will be kept which  has not been published.}
}
\value{
A horizontal bar plot for colocalisation of information.
}
\description{
Generates a plot of the data obtained from colocalisationsForGene
function results.
}
\examples{
\dontrun{
plot_out <- colocalisationsForGene(genes = "ENSG00000169174") \%>\% otargen::plot_coloc(biobank = FALSE)
plot_out <- colocalisationsForGene(genes = "PCSK9") \%>\% otargen::plot_coloc(biobank = TRUE)
}
}
