% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_manhattan.R
\name{plot_manhattan}
\alias{plot_manhattan}
\title{Plot the data obtained from the manhattan associations.}
\usage{
plot_manhattan(data)
}
\arguments{
\item{\emph{data}}{Data frame: result of the manhattan function containing SNP information}
}
\value{
A manhattan plot representing SNPs on the x-axis with the corresponding
negative logarithm of p-value on the y-axis.
}
\description{
The plot is generated using the p-value and variant information obtained from
the manhattan function. Users can provide p-value as a parameter to obtain a plot with the significant SNPs.
}
\examples{
\dontrun{
otargen::manhattan(studyid = "GCST003044") \%>\% otargen::plot_manhattan()
}
}
