% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_phewas.R
\name{plot_phewas}
\alias{plot_phewas}
\title{Plot the PheWAS function results}
\usage{
plot_phewas(
  data,
  disease = TRUE,
  source = c("GCST", "FINNGEN", "NEALE", "SAIGE")
)
}
\arguments{
\item{data}{is the result of PheWAS function in data frame format, contacting the phewas information for a variand id}

\item{disease}{is a logical TRUE and FALSE variable, with the default value of TRUE to filter the phewas data for disease.}

\item{source}{is a character vector of choices for data sources of phewas analyis including FINNGEN, GCST, NEAL (UKbioBANK), and SAGE}

\item{...}{}
}
\description{
Plot the PheWAS function results
}
\examples{
pheWAS(variantid = "14_87978408_G_A") \%>\% plot_phewas(disease = TRUE)

}
