% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qtlColocalisationVariantQuery.R
\name{qtlColocalisationVariantQuery}
\alias{qtlColocalisationVariantQuery}
\title{Get Colocalization data for a variant in specific study}
\usage{
qtlColocalisationVariantQuery(studyid, variantid)
}
\arguments{
\item{studyid}{is the Open Target Genetics generated id for gwas studies.}

\item{variantid}{is Open Target Genetics generated id for each variant in the databse.}
}
\value{
a data frame of the colocalization information for a lead variant
}
\description{
Get Colocalization data for a variant in specific study
}
