% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qtlCredibleSet.R
\name{qtlCredibleSet}
\alias{qtlCredibleSet}
\title{Retrieves QTL credible set data}
\usage{
qtlCredibleSet(studyid, variantid, gene, biofeature)
}
\arguments{
\item{\emph{studyid}}{String: Open Targets Genetics generated ID for GWAS study.}

\item{\emph{variantid}}{String: Open Targets Genetics generated ID for variant (CHRPOSITION_REFALLELE_ALTALLELE or rsId).}

\item{\emph{gene}}{String: Gene ENSEMBL ID or gene name.}

\item{\emph{biofeature}}{String: Represents either a tissue, cell type, aggregation type, protein type, etc.}
}
\value{
Returns a data frame of results from the QTL credible set of variants consisting of the following columns:
\itemize{
  \item{\code{tagVariant.id}:} \emph{Character vector}. Tag variant ID.
  \item{\code{tagVariant.rsId}:} \emph{Character vector}. Tag variant rsID.
  \item{\code{pval}:} \emph{Numeric}. P-value.
  \item{\code{se}:} \emph{Numeric}. Standard error.
  \item{\code{beta}:} \emph{Numeric}. Beta value.
  \item{\code{postProb}:} \emph{Numeric}. Posterior probability.
  \item{\code{MultisignalMethod}:} \emph{Character vector}. Multisignal method.
  \item{\code{logABF}:} \emph{Numeric}. Logarithm of approximate Bayes factor.
  \item{\code{is95}:} \emph{Logical}. Indicates if the variant has a 95% confidence.
  \item{\code{is99}:} \emph{Logical}. Indicates if the variant has a 99% confidence.
}
}
\description{
For a specific gene, lead variant, study ID, and biofeature, the function obtains tag variant information
that are considered to have a high probability of being truly associated with the trait and the corresponding scores.
}
\examples{
\dontrun{
result <- qtlCredibleSet(studyid = "Braineac2", variantid = "1_55053079_C_T",
    gene = "ENSG00000169174", biofeature = "SUBSTANTIA_NIGRA")
result <- qtlCredibleSet(studyid = "Braineac2", variantid = "rs7552841",
    gene = "PCSK9", biofeature = "SUBSTANTIA_NIGRA")
}

}
