% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_custom_query.R
\name{run_custom_query}
\alias{run_custom_query}
\title{Running custom GraphQL queries}
\usage{
run_custom_query(variableList, query, query_name)
}
\arguments{
\item{variableList}{is a list format which includes the key value pair list of genes/variants/study ids to be queries.}

\item{query}{is a GraphQL desired query body to be run.}

\item{query_name}{is a string format of the query name}
}
\value{
a flatten json file format
}
\description{
Running custom GraphQL queries
}
\examples{
\dontrun{
otargen::run_custom_query (variableList, query, query_name)
}

}
