% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/safetyQuery.R
\name{safetyQuery}
\alias{safetyQuery}
\title{Retrieve Safety Liabilities data for a specified gene.}
\usage{
safetyQuery(ensgId)
}
\arguments{
\item{ensgId}{Character: ENSEMBL ID of the target gene (e.g., ENSG00000141510).}
}
\value{
Returns a tibble containing safety liabilities data for the specified gene.
}
\description{
This function queries the Open Targets GraphQL API to retrieve safety liabilities data
for a specified gene.
}
\examples{
\dontrun{
result <- safetyQuery(ensgId = "ENSG00000141510")
}
}
