% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/studiesAndLeadVariantsForGeneByL2G.R
\name{studiesAndLeadVariantsForGeneByL2G}
\alias{studiesAndLeadVariantsForGeneByL2G}
\title{Query L2G model summary data for a gene}
\usage{
studiesAndLeadVariantsForGeneByL2G(ensmbl_ids)
}
\arguments{
\item{ensembl}{id is a identification id for genes by ensembl database}
}
\value{
a dataframe including the queried gene indentity and its colocalization
data
}
\description{
Query L2G model summary data for a gene
}
