% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/studyInfo.R
\name{studyInfo}
\alias{studyInfo}
\title{Retrieves GWAS study id information.}
\usage{
studyInfo(studyid)
}
\arguments{
\item{\emph{studyid}}{String: Open Targets Genetics generated ID for a GWAS study.}
}
\value{
Returns a tibble data table containing the GWAS study information. The table consists of the following columns:
\itemize{
  \item{\code{studyId}:} \emph{Character}. Study ID.
  \item{\code{traitReported}:} \emph{Character}. Reported trait.
  \item{\code{source}:} \emph{Character}. Source.
  \item{\code{traitEfos}:} \emph{Character}. Trait EFO ID.
  \item{\code{pmid}:} \emph{Character}. PubMed ID.
  \item{\code{pubDate}:} \emph{Character}. Publication date.
  \item{\code{pubJournal}:} \emph{Character}. Publication journal.
  \item{\code{pubTitle}:} \emph{Character}. Publication title.
  \item{\code{pubAuthor}:} \emph{Character}. Publication author.
  \item{\code{hasSumstats}:} \emph{Character}. Indicates if the study has summary statistics.
  \item{\code{ancestryInitial}:} \emph{Character}. Initial ancestry.
  \item{\code{nInitial}:} \emph{Character}. Initial sample size.
  \item{\code{nReplication}:} \emph{Character}. Replication sample size.
  \item{\code{traitCategory}:} \emph{Character}. Trait category.
  \item{\code{numAssocLoci}:} \emph{Character}. Number of associated loci.
  \item{\code{nTotal}:} \emph{Character}. Total sample size.
}
}
\description{
For a given study id, this function returns all the relevant information about
a GWAS study, such as PubMed ID, studied trait EFO ID, case/control size, etc.
}
\examples{
\dontrun{
result <- studyInfo(studyid = "GCST90002357")
}
}
