% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tagVariantsAndStudiesForIndexVariant.R
\name{tagVariantsAndStudiesForIndexVariant}
\alias{tagVariantsAndStudiesForIndexVariant}
\title{Gives out tagVariants and studies for a given index variant.}
\usage{
tagVariantsAndStudiesForIndexVariant(variantid, pageindex = 0, pagesize = 20)
}
\arguments{
\item{variantid}{is the Open Target Genetics generated id for each variants.}

\item{pageIndex}{pagination index >= 0. Index of the current page.}

\item{pageSize}{pagination size > 0. No. of records in a page. Default: 20}
}
\value{
A dataframe containing the variant associations connnected to the input index variant.
}
\description{
Gives out tagVariants and studies for a given index variant.
}
\examples{
tagVariantsAndStudiesForIndexVariant("GCST006614_3")
tagVariantsAndStudiesForIndexVariant("NEALE2_6177_1", pageindex=1, pagesize=50)
}
