% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topOverlappedStudies.R
\name{topOverlappedStudies}
\alias{topOverlappedStudies}
\title{Retrieves the top studies having overlap with the given input study.}
\usage{
topOverlappedStudies(studyid, pageindex = 0, pagesize = 20)
}
\arguments{
\item{\emph{studyid}}{String: Open Target Genetics generated id for GWAS study.}

\item{\emph{pageindex}}{Int: Index of the current page, pagination index >= 0.}

\item{\emph{pagesize}}{Int: No. of records in a page, pagination size > 0.}
}
\value{
Returns a data frame with top studies containing the following columns:

\enumerate{
\item study.studyId
\item study.traitReported
\item study.traitCategory
\item topStudiesByLociOverlap.studyId
\item topStudiesByLociOverlap.study.studyId
\item topStudiesByLociOverlap.study.traitReported
\item topStudiesByLociOverlap.study.traitCategory
\item topStudiesByLociOverlap.numOverlapLoci
}
}
\description{
Fetches the top associated studies for a given study id ordered by loci overlap.
}
\examples{
\dontrun{
otargen::topOverlappedStudies(studyid = "GCST006614_3")
otargen::topOverlappedStudies(studyid = "NEALE2_6177_1", pageindex = 1, pagesize = 50)
}
}
