% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniProtVariantsQuery.R
\name{uniProtVariantsQuery}
\alias{uniProtVariantsQuery}
\title{Retrieve UniProt Variants data for a specified variant.}
\usage{
uniProtVariantsQuery(variantId)
}
\arguments{
\item{variantId}{Character: ID of the target variant (e.g., "4_1804392_G_A").}
}
\value{
Returns a tibble containing UniProt variants data for the specified variant.
}
\description{
This function queries the Open Targets GraphQL API to retrieve UniProt variants data
for a specified variant.
}
\examples{
\dontrun{
result <- uniProtVariantsQuery(variantId = "4_1804392_G_A")
}
}
