% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variantEffectPredictorQuery.R
\name{variantEffectPredictorQuery}
\alias{variantEffectPredictorQuery}
\title{Retrieve Variant Effect Predictor data for a specified variant.}
\usage{
variantEffectPredictorQuery(variantId)
}
\arguments{
\item{variantId}{Character: ID of the target variant (e.g., "4_1804392_G_A").}
}
\value{
Returns a tibble containing variant effect predictor data for the specified variant.
}
\description{
This function queries the Open Targets GraphQL API to retrieve variant effect predictor data
for a specified variant.
}
\examples{
\dontrun{
result <- variantEffectPredictorQuery(variantId = "4_1804392_G_A")
}
}
