% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variantInfo.R
\name{variantInfo}
\alias{variantInfo}
\title{Retrieves the information about the input variant id.}
\usage{
variantInfo(variantid)
}
\arguments{
\item{variantid}{String: Open Target Genetics generated id for variant (CHR_POSITION_REFALLELE_ALT_ALLELE or rsId).}
}
\value{
Returns a data frame containing the variant information with the following columns:

\enumerate{
\item chromosome
\item position
\item refAllele
\item altAllele
\item rsId
\item chromosomeB37
\item positionB37
\item id
\item nearestGene.id
\item nearestGene.symbol
\item nearestGeneDistance
\item nearestCodingGene.id
\item nearestCodingGene.symbol
\item nearestCodingGeneDistance
\item mostSevereConsequence
\item caddRaw
\item caddPhred
\item gnomadAFR
\item gnomadAMR
\item gnomadASJ
\item gnomadEAS
\item gnomadFIN
\item gnomadNFE
\item gnomadNFEEST
\item gnomadNFENWE
}
}
\description{
Retrieves the information about the input variant id.
}
\examples{
\dontrun{
otargen::variantInfo(variantid = "1_55039974_G_T")
otargen::variantInfo(variantid = "rs11591147")
}
}
