% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/associatedTargets.R
\name{associatedTargets}
\alias{associatedTargets}
\title{Get associated targets for a disease category}
\usage{
associatedTargets(efoid, n)
}
\arguments{
\item{efoid}{A string. The Experimental Factor Ontology(EFO) id}

\item{n}{An integer. Number of the record to return}
}
\value{
Returns a dataframe including a data table of the associated drug targets
}
\description{
Retrive specified number of drug targets (n) associated with a disease with specified EFO id
}
