% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colocalisationsForGene.R
\name{colocalisationsForGene}
\alias{colocalisationsForGene}
\title{Retrieves colocalisation data for a gene.}
\usage{
colocalisationsForGene(genes)
}
\arguments{
\item{\emph{genes}}{String: one or more gene ENSEMBL identifier or gene name.}
}
\value{
Returns a tibble including the query gene(s) colocalisation data.

The output data frame contain the following columns:

\enumerate{
\item Study
\item Trait_reported
\item Lead_variant
\item Molecular_trait
\item Gene_symbol
\item Tissue
\item Source
\item H3
\item log2(H4/H3)
\item Title
\item Author
\item Has_sumstats
\item numAssocLoci
\item nInitial cohort
\item study_nReplication
\item study_nCases
\item Publication_date
\item Journal
\item Pubmed_id
}
}
\description{
This function retrieves colocalization data for input ENSEMBL gene IDs or a gene name and
it supports multiple gene IDs as a list of inputs. However, providing a limited number of
genes is recommended to prevent excessive traffic on the OTG servers and ensure optimal
performance.The return is a data table (tibble format) of studies that have evidence of colocalisation
 with molecular QTLs for given gene(s).
}
\examples{
\dontrun{
otargen::colocalisationsForGene(genes=list("ENSG00000163946",
"ENSG00000169174", "ENSG00000143001"))
otargen::colocalisationsForGene(genes="ENSG00000169174")
otargen::colocalisationsForGene(genes=list("TP53", "TASOR"))
otargen::colocalisationsForGene(genes="TP53")
}
}
