% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneInfo.R
\name{geneInfo}
\alias{geneInfo}
\title{Get gene information for an input ENSEMBL identifier}
\usage{
geneInfo(geneid)
}
\arguments{
\item{geneid}{An ENSEMBL identifier for a gene}
}
\value{
A dataframe containing the input gene information like symbol, chromosome information, etc.
}
\description{
Get gene information for an input ENSEMBL identifier
}
\examples{
geneInfo("ENSG00000169174")
}
