% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genesForVariant.R
\name{genesForVariant}
\alias{genesForVariant}
\title{Fetches variant-to-gene information with individual QTL associations, intervals, distances, and functional predictions.}
\usage{
genesForVariant(variantid)
}
\arguments{
\item{variantId}{A character string specifying the ID of the variant for which to fetch gene information.}
}
\value{
A list with the following components:
\itemize{
  \item \code{v2g}: Is a data frame with all variant-to-gene information with the following data structure.
    - \code{gene.symbol}: character
    - \code{variant}: character
    - \code{overallScore}: numeric
    - \code{gene.id}: character

  \item \code{tssd}: Is a data frame with all details on colocalization scores effect of variant
   in expression of the genes across tissues with the following data structure:
    - \code{gene.symbol}: character
    - \code{variant}: character
    - \code{typeId}: character
    - \code{sourceId}: character
    - \code{aggregatedScore}: numeric
    - \code{tissues_distance}: integer
    - \code{tissues_score}: numeric
    - \code{tissues_quantile}: numeric
    - \code{tissues_id}: character
    - \code{tissues_name}: character

  \item \code{qtls}: List of QTL associations between genes and variants across analysed tissues with the
   following data structure:
    - \code{gene.symbol}: character
    - \code{variant}: character
    - \code{typeId}: character
    - \code{aggregatedScore}: numeric
    - \code{tissues_quantile}: numeric
    - \code{tissues_beta}: numeric
    - \code{tissues_pval}: numeric
    - \code{tissues_id}: character
    - \code{tissues_name}: character

  \item \code{chromatin}: Is a data frame which includes all information on chromatin interactions effect
  involving genes and variants with the following data structure:
    - \code{gene.symbol}: character
    - \code{variant}: character
    - \code{typeId}: character
    - \code{sourceId}: character
    - \code{aggregatedScore}: numeric
    - \code{tissues_quantile}: numeric
    - \code{tissues_score}: numeric
    - \code{tissues_id}: character
    - \code{tissues_name}: character

  \item \code{functionalpred}: Is a data frame that includes predicted functional effect of
   variants on genes across tissues with the followint data structure:
    - \code{gene.symbol}: character
    - \code{variant}: character
    - \code{typeId}: character
    - \code{sourceId}: character
    - \code{aggregatedScore}: numeric
    - \code{tissues_maxEffectLabel}: character
    - \code{tissues_maxEffectScore}: numeric
    - \code{tissues_id}: character
    - \code{tissues_name}: character
}
}
\description{
This function retrieves various information related to variants and genes, including QTL associations, distances, chromatin interactions, and functional predictions.
}
\examples{
\dontrun{
  result <- genesForVariant(variantId = "1_154453788_C_T")
  print(result)

}
}
