% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genesForVariant.R
\name{genesForVariant}
\alias{genesForVariant}
\title{Retrieves variant to Gene (V2G) data.}
\usage{
genesForVariant(variantid)
}
\arguments{
\item{variantid}{String: Open Target Genetics generated id for variant (CHR_POSITION_REFALLELE_ALT_ALLELE or rsId).}
}
\value{
A list of tibble data tables containing variant to gene (v2g) information with
individual QTL associations, intervals, distances and functional predictions.

v2g table consists of following columns:

\enumerate{
\item gene.symbol
\item variant
\item overallScore
\item gene.id
}

tssd (distance to TSS) table consists of following columns:

\enumerate{
\item gene.symbol
\item variant
\item typeId
\item sourceId
\item aggregatedScore
\item tissues_distance
\item tissues_score
\item tissues_quantile
\item tissues_id
\item tissues_name
}

qtl table consists of the following columns:

\enumerate{
\item gene.symbol
\item variant
\item typeId
\item aggregatedScore
\item tissues_quantile
\item tissues_beta
\item tissues_pval
\item tissues_id
\item tissues_name
}

chromatin table consists of the following columns:

\enumerate{
\item gene.symbol
\item variant
\item typeId
\item sourceId
\item aggregatedScore
\item tissues_quantile
\item tissues_score
\item tissues_id
\item tissues_name
}

functional prediction (functionalpred) table consists of following columns:

\enumerate{
\item gene.symbol
\item variant
\item typeId
\item sourceId
\item aggregatedScore
\item tissues_maxEffectLabel
\item tissues_maxEffectScore
\item tissues_id
\item tissues_name
}
}
\description{
This function gets variant ID formatted as CHR_POSITION_REF, ALLELE_ALT_ALLELE or canonical SNPs ID as an input and returns a list of
tibble data tables including ranked gene table and corresponding functional genomics data tables that are functionally implicated by this variant.
}
\examples{
\dontrun{
otargen::genesForVariant(variantid = "1_154453788_C_T")
otargen::genesForVariant(variantid = "rs4129267")
}
}
