% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwasColocalisation.R
\name{gwasColocalisation}
\alias{gwasColocalisation}
\title{Retrieves GWAS colocalisation data.}
\usage{
gwasColocalisation(studyid, variantid)
}
\arguments{
\item{studyid}{String: Open Target Genetics generated id for GWAS study.}

\item{variantid}{String: Open Target Genetics generated id for variant (CHR_POSITION_REFALLELE_ALT_ALLELE or rsId).}
}
\value{
Returns a data frame of the studies which colocalise with the input variant and study. The table consists of the following columns:

\enumerate{
\item study.studyId
\item study.traitReported
\item study.traitCategory
\item indexVariant.id
\item indexVariant.position
\item indexVariant.chromosome
\item indexVariant.rsId
\item beta
\item h3
\item h4
\item log2h4h3
}
}
\description{
Provided with study id and variant id, this function returns a tibble data table of all
colocalised index variants with the corresponding GWAS study information and colocalisation scores.
}
\examples{
\dontrun{
otargen::gwasColocalisation(studyid = "GCST90002357", variantid = "1_154119580_C_A")
otargen::gwasColocalisation(studyid = "GCST90002357", variantid = "rs2494663")
}
}
