% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwasColocalisationForRegion.R
\name{gwasColocalisationForRegion}
\alias{gwasColocalisationForRegion}
\title{Get gwas colocalisation data for a given region}
\usage{
gwasColocalisationForRegion(chromosome, start, end)
}
\arguments{
\item{chromosome}{chromosome number given as string.}

\item{start}{start position of the specified chromosome.}

\item{end}{end position of the specified chromosome.}
}
\value{
A data frame with gwas colocalisation data containing studies and variants for the queried chromosome and region
}
\description{
Get gwas colocalisation data for a given region
}
\examples{
gwasColocalisationForRegion("1", 153992685, 154155116)
}
