% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwasColocalisationForRegion.R
\name{gwasColocalisationForRegion}
\alias{gwasColocalisationForRegion}
\title{Retrieves GWAS colocalisation data for a region.}
\usage{
gwasColocalisationForRegion(chromosome, start, end)
}
\arguments{
\item{\emph{chromosome}}{String: chromosome number as string.}

\item{\emph{start}}{Long: start position of the specified chromosome.}

\item{\emph{end}}{Long: end position of the specified chromosome.}
}
\value{
Returns a data frame with the following columns:

\enumerate{
\item study.studyId
\item study.traitReported
\item study.traitCategory
\item indexVariant.id
\item indexVariant.position
\item indexVariant.chromosome
\item indexVariant.rsId
\item beta
\item h3
\item h4
\item log2h4h3
}
}
\description{
With providing a defined chromosomal right this function reruns a tibble data table
of colocalised variants to the left and right side of the provided regions including
the calculated colocalization scores.
}
\examples{
\dontrun{
otargen::gwasColocalisationForRegion(chromosome = "1", start = 153992685, end = 154155116)
}
}
