% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwasCredibleSet.R
\name{gwasCredibleSet}
\alias{gwasCredibleSet}
\title{Get GWAS credible set data for a variant in a study}
\usage{
gwasCredibleSet(studyid, variantid)
}
\arguments{
\item{\emph{studyid}}{String: Open Target Genetics generated id for GWAS study.}

\item{\emph{variantid}}{String: Open Target Genetics generated id for variant (CHRPOSITION_REFALLELE_ALTALLELE or rsId).}
}
\value{
Returns a data frame of results from credible set of variants for a specific lead variant with the following columns:

\enumerate{
\item tagVariant.id
\item tagVariant.rsId
\item beta
\item postProb
\item pval
\item se
\item MultisignalMethod
\item logABF
\item is95
\item is99
}
}
\description{
Provided with study id and a lead variant id parameters,
this functions returns a data table consisting of all the associated credible
set tag variants with the corresponding statistical data.
}
\examples{
\dontrun{
otargen::gwasCredibleSet(studyid="GCST90002357", variantid="1_154119580_C_A")
otargen::gwasCredibleSet(studyid="GCST90002357", variantid="rs2494663")
}
}
