% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwasCredibleSet.R
\name{gwasCredibleSet}
\alias{gwasCredibleSet}
\title{Get gwas credible set data for a variant in a study}
\usage{
gwasCredibleSet(studyid, variantid)
}
\arguments{
\item{studyid}{is the Open Target Genetics generated id for gwas studies.}

\item{variantid}{is Open Target Genetics generated id for each variant in the databse.}
}
\value{
A data frame of results from credible set of variants for a specific lead variant.
}
\description{
Get gwas credible set data for a variant in a study
}
\examples{
gwasCredibleSet("GCST90002357", "1_154119580_C_A")
}
