% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwasRegional.R
\name{gwasRegional}
\alias{gwasRegional}
\title{GWAS Regional Association of a study.}
\usage{
gwasRegional(studyid, chromosome, start, end)
}
\arguments{
\item{\emph{studyid}}{String: Open Target Genetics generated id for GWAS study.}

\item{\emph{chromosome}}{String: chromosome number as string.}

\item{\emph{start}}{Long: start position of the specified chromosome.}

\item{\emph{end}}{Long: end position of the specified chromosome.}
}
\value{
Returns a data table of variant information and pvalue. The table consists of the following columns:

\enumerate{
\item variant.id
\item variant.chromosome
\item variant.position
\item pval
}
}
\description{
For a given study id and a chromosomal region information,
this function returns a data table in a tibble format with all variants and
their respective p-values.
}
\examples{
\dontrun{
otargen::gwasRegional(studyid="GCST90002357", chromosome="1", start=153992685, end=154155116)
}
}
