% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwasRegional.R
\name{gwasRegional}
\alias{gwasRegional}
\title{GWAS Regional Association of a study.}
\usage{
gwasRegional(studyid, chromosome, start, end)
}
\arguments{
\item{studyid}{is the Open Target Genetics generated id for gwas studies.}

\item{chromosome}{chromosome number as string.}

\item{start}{start position of the specified chromosome.}

\item{end}{end position of the specified chromosome.}
}
\value{
A dataframe of variants and pval score which have regional association with the given study id, chromosome number and the specified region.
}
\description{
GWAS Regional Association of a study.
}
\examples{
gwasRegional("GCST90002357", "1", 153992685, 154155116)
}
