% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/literatureDrugQuery.R
\name{literatureDrugQuery}
\alias{literatureDrugQuery}
\title{Collect litreture evidance for known drug for a disease}
\usage{
literatureDrugQuery(efoid)
}
\arguments{
\item{efoid}{is experimental factor ontology}
}
\value{
a dataframe including the queried gene indentity and its colocalization data
}
\description{
Collect litreture evidance for known drug for a disease
}
