% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manhattan.R
\name{manhattan}
\alias{manhattan}
\title{Manhattan association for a given study.}
\usage{
manhattan(studyid, pageindex = 0, pagesize = 20)
}
\arguments{
\item{studyid}{is the Open Target Genetics generated id for gwas studies.}

\item{pageIndex}{pageIndex pagination index >= 0. Index of the current page.}

\item{pageSize}{pagination size > 0. No. of records in a page. Default: 20}
}
\value{
A dataframe containing manhattan associations for the input study id given.
}
\description{
Manhattan association for a given study.
}
\examples{
manhattan("GCST90002357")
manhattan("GCST90002357", 2, 50)
}
