% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manhattan.R
\name{manhattan}
\alias{manhattan}
\title{Manhattan association for a given study.}
\usage{
manhattan(studyid, pageindex = 0, pagesize = 100)
}
\arguments{
\item{\emph{studyid}}{String: Open Targets Genetics generated ID for GWAS study.}

\item{\emph{pageindex}}{Int: Index of the current page, pagination index >= 0.}

\item{\emph{pagesize}}{Int: Number of records in a page, pagination size > 0.}
}
\value{
Returns a data frame containing Manhattan associations for the input study ID. The table consists of the following columns:
\itemize{
  \item{\code{pval_mantissa}:} \emph{Numeric vector}. Mantissa of the p-value.
  \item{\code{pval_exponent}:} \emph{Integer vector}. Exponent of the p-value.
  \item{\code{credible_set_size}:} \emph{Integer vector}. Size of the credible set.
  \item{\code{ld_set_size}:} \emph{Integer vector}. Size of the LD set.
  \item{\code{total_set_size}:} \emph{Integer vector}. Total size of the set.
  \item{\code{pval}:} \emph{Numeric vector}. P-value.
  \item{\code{odds_ratio}:} \emph{Logical vector}. Odds ratio.
  \item{\code{odds_ratio_ci_lower}:} \emph{Logical vector}. Lower confidence interval of the odds ratio.
  \item{\code{odds_ratio_ci_upper}:} \emph{Logical vector}. Upper confidence interval of the odds ratio.
  \item{\code{beta}:} \emph{Numeric vector}. Beta value.
  \item{\code{beta_ci_lower}:} \emph{Numeric vector}. Lower confidence interval of the beta value.
  \item{\code{beta_ci_upper}:} \emph{Numeric vector}. Upper confidence interval of the beta value.
  \item{\code{direction}:} \emph{Character vector}. Direction of the effect.
  \item{\code{best_genes_score}:} \emph{Numeric vector}. Score of the best genes.
  \item{\code{best_genes_gene_id}:} \emph{Character vector}. Gene ID of the best genes.
  \item{\code{best_genes_gene_symbol}:} \emph{Character vector}. Gene symbol of the best genes.
  \item{\code{best_coloc_genes_score}:} \emph{Numeric vector}. Score of the best colocated genes.
  \item{\code{best_coloc_genes_gene_id}:} \emph{Character vector}. Gene ID of the best colocated genes.
  \item{\code{best_coloc_genes_gene_symbol}:} \emph{Character vector}. Gene symbol of the best colocated genes.
  \item{\code{best_locus2genes_score}:} \emph{Numeric vector}. Score of the best locus to genes.
  \item{\code{best_locus2genes_gene_id}:} \emph{Character vector}. Gene ID of the best locus to genes.
  \item{\code{best_locus2genes_gene_symbol}:} \emph{Character vector}. Gene symbol of the best locus to genes.
  \item{\code{variant_id}:} \emph{Character vector}. Variant ID.
  \item{\code{variant_position}:} \emph{Integer vector}. Variant position.
  \item{\code{variant_chromosome}:} \emph{Character vector}. Variant chromosome.
  \item{\code{variant_rs_id}:} \emph{Character vector}. Variant rsID.
}
}
\description{
The data generated is used to plot the Manhattan association. Manhattan association is used to graphically represent genetic association studies, particularly in genome-wide association studies (GWAS). It displays the results of statistical associations between genetic variants and a trait or disease of interest across the genome.
}
\examples{
\dontrun{
result <- manhattan(studyid = "GCST90002357")
result <- manhattan(studyid = "GCST90002357", pageindex = 2, pagesize = 50)
}
}
