% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manhattan.R
\name{manhattan}
\alias{manhattan}
\title{Manhattan association for a given study.}
\usage{
manhattan(studyid, pageindex = 0, pagesize = 100)
}
\arguments{
\item{\emph{studyid}}{String: Open Targets Genetics generated id for GWAS study.}

\item{\emph{pageindex}}{Int: Index of the current page, pagination index >= 0.}

\item{\emph{pagesize}}{Int: No. of records in a page, pagination size > 0.}
}
\value{
Returns a data frame containing manhattan associations for the input study id given. The table consists of the following columns:

\enumerate{
\item pval_mantissa
\item pval_exponent
\item credible_set_size
\item ld_set_size
\item total_set_size
\item pval
\item odds_ratio
\item odds_ratio_ci_lower
\item odds_ratio_ci_upper
\item beta
\item beta_ci_lower
\item beta_ci_upper
\item direction
\item best_genes_score
\item best_genes_gene_id
\item best_genes_gene_symbol
\item best_coloc_genes_score
\item best_coloc_genes_gene_id
\item best_coloc_genes_gene_symbol
\item best_locus2genes_score
\item best_locus2genes_gene_id
\item best_locus2genes_gene_symbol
\item variant_id
\item variant_position
\item variant_chromosome
\item variant_rs_id
}
}
\description{
The data generated is used to plot the Manhattan association. Manhattan association is used to  graphically representat genetic association studies,
particularly in genome-wide association studies (GWAS). It displays the results of statistical associations between genetic variants and a trait or
disease of interest across the genome.
}
\examples{
\dontrun{
otargen::manhattan(studyid = "GCST90002357")
otargen::manhattan(studyid = "GCST90002357", pageindex = 2, pagesize = 50)
}
}
