% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlapInfoForStudy.R
\name{overlapInfoForStudy}
\alias{overlapInfoForStudy}
\title{Get overlap info for a study with given list of studies}
\usage{
overlapInfoForStudy(studyid, studyids = list())
}
\arguments{
\item{studyid}{is the Open Target Genetics generated id for gwas studies.}

\item{studyids}{is a list of Open Target Genetics generated id for gwas studies.}
}
\value{
A data frame of results containing the overlap info between the input studyid and the list of studyids.
}
\description{
Get overlap info for a study with given list of studies
}
\examples{
overlapInfoForStudy("GCST90002357", list("GCST90025975","GCST90025962"))
}
