% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlapInfoForStudy.R
\name{overlapInfoForStudy}
\alias{overlapInfoForStudy}
\title{Retrieves overlap info for a study and a list of studies}
\usage{
overlapInfoForStudy(studyid, studyids = list())
}
\arguments{
\item{\emph{studyid}}{String: Open Target Genetics generated id for GWAS study.}

\item{\emph{studyids}}{List of Strings: list of Open Target Genetics generated id for GWAS studies.}
}
\value{
A list containing a data frame of overlap information and the variant intersection set.

The overlap information table (overlap_info) consists of the following columns:
\enumerate{
\item studyId
\item traitReported
\item traitCategory
\item variantIdA
\item variantIdB
\item overlapAB
\item distinctA
\item distinctB
\item study.studyId
\item study.traitReported
\item study.traitCategory
}
}
\description{
For an input study id and a list of other study ids, this function returns two elements.
One contains the overlap information in a table format. The other element is the variant intersection set.
It represents an overlap between two variants of the two given studies.
}
\examples{
\dontrun{
otargen::overlapInfoForStudy(studyid="GCST90002357", studyids=list("GCST90025975","GCST90025962"))
}
}
