% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pheWAS.R
\name{pheWAS}
\alias{pheWAS}
\title{Phenome Wide Association Studies}
\usage{
pheWAS(variantid)
}
\arguments{
\item{variantid}{is the Open Target Genetics generated id for each variants}
}
\value{
Returns a data frame with PheWAS associations consisting of the following columns:

\enumerate{
\item totalGWASStudies
\item pval
\item beta
\item oddsRatio
\item nTotal
\item study.studyId
\item study.source
\item study.pmid
\item study.pubDate
\item study.traitReported
\item study.traitCategory
}
}
\description{
Fetches the phenome-wide association study for a variant.
}
\examples{
\dontrun{
otargen::pheWAS(variantid ="1_154549918_C_A")
otargen::pheWAS(variantid ="rs72698179")
}
}
