% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pheWAS.R
\name{pheWAS}
\alias{pheWAS}
\title{Phenome Wide Association Studies}
\usage{
pheWAS(variantid)
}
\arguments{
\item{variantid}{is the Open Target Genetics generated id for each variants}
}
\value{
A data frame with PheWAS associations in it
}
\description{
Phenome Wide Association Studies
}
\examples{
pheWAS("1_154549918_C_A")
pheWAS("1_154562290_G_A")
}
