% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pheWAS.R
\name{pheWAS}
\alias{pheWAS}
\title{Retrieves Phenome Wide Association Studies for a variant.}
\usage{
pheWAS(variantid)
}
\arguments{
\item{variantid}{is the Open Target Genetics generated id for each variants}
}
\value{
Returns a data frame with PheWAS associations consisting of the following columns:

\enumerate{
\item totalGWASStudies
\item pval
\item beta
\item oddsRatio
\item nTotal
\item study.studyId
\item study.source
\item study.pmid
\item study.pubDate
\item study.traitReported
\item study.traitCategory
}
}
\description{
PheWAS (Phenome-wide association study) is a method that investigates the relationships
between genetic variants and a traits or phenotypes. It helps in the study of their
potential influence on multiple traits or diseases concurrently.
}
\examples{
\dontrun{
otargen::pheWAS(variantid ="1_154549918_C_A")
otargen::pheWAS(variantid ="rs72698179")
}
}
