% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_l2g.R, R/plot_manhattan.R
\name{plot_l2g}
\alias{plot_l2g}
\title{Plot the the scores obtained from the L2G model results}
\usage{
plot_l2g(data, pvalue = 1e-07)

plot_l2g(data, pvalue = 1e-07)
}
\arguments{
\item{data}{Data frame: result of the manhattan function containing SNP information}

\item{pvalue}{Float: pvalue cut off.}

\item{disease_efo}{is the input efo id to filter the L2G data for a particular disease.}
}
\value{
A radar plot for the input disease and the genes associated with that disease. The variables shown include L2G score, chromatin interaction, variant
pathogenicity and distance.

A manhattan plot representing SNPs on the x-axis with the corresponding
negative logarithm of p-value on the y-axis.
}
\description{
The plot is generated using the p-value and variant information obtained from
the manhattan function. Users can provide p-value as a parameter to obtain a plot with the significant SNPs.
}
\examples{
studiesAndLeadVariantsForGeneByL2G(list("ENSG00000167207","ENSG00000096968","ENSG00000138821", "ENSG00000125255")) \%>\% plot_l2g(disease = "EFO_0003767")

manhattan(studyid = "GCST003044") \%>\% plot_manhattan(pval= 10e-8)

}
