% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_l2g.R
\name{plot_l2g}
\alias{plot_l2g}
\title{Plot the the scores obtained from the L2G model results}
\usage{
plot_l2g(data, disease_efo = NULL)
}
\arguments{
\item{data}{Data frame: result of studiesAndLeadVariantsForGeneByL2G function.}

\item{disease_efo}{String: input efo id to filter the L2G data for a particular disease.}
}
\value{
A radar plot for the input disease and the genes associated with that disease. The variables shown include L2G score, chromatin interaction, variant
pathogenicity and distance.
}
\description{
Plot the the scores obtained from the L2G model results
}
\examples{
\dontrun{
studiesAndLeadVariantsForGeneByL2G(list("ENSG00000167207",
"ENSG00000096968","ENSG00000138821", "ENSG00000125255")) \%>\%
otargen::plot_l2g(disease = "EFO_0003767")
}

}
