% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_phewas.R
\name{plot_phewas}
\alias{plot_phewas}
\title{Plot the PheWAS function results}
\usage{
plot_phewas(
  data,
  disease = TRUE,
  source = c("GCST", "FINNGEN", "NEALE", "SAIGE")
)
}
\arguments{
\item{\emph{data}}{Data Frame: result of PheWAS function in data frame format, contacting the PheWAS information for a variand id}

\item{\emph{disease}}{Logical: a logical TRUE and FALSE variable, with the default value of TRUE to filter the PheWAS data for disease.}

\item{\emph{source}}{Character vector: choices for data sources of PheWAS analysis including FINNGEN, GCST, NEAL (UKBioBANK), and SAGE.}
}
\value{
A plot to prioritize variants based on their -log10(pval).
}
\description{
Phenome wide association study results are filtered for diseases and the -log10 p-value is plotted against the traits.
It presents the association results between genetic variants and multiple traits. The traits with p-value < 0.005 is labelled.
}
\examples{
\dontrun{
otargen::pheWAS(variantid = "14_87978408_G_A") \%>\% otargen::plot_phewas(disease = TRUE)
}
}
