% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qtlColocalisationVariantQuery.R
\name{qtlColocalisationVariantQuery}
\alias{qtlColocalisationVariantQuery}
\title{Retrieves colocalisation data for a variant in a study}
\usage{
qtlColocalisationVariantQuery(studyid, variantid)
}
\arguments{
\item{studyid}{String: Open Target Genetics generated id for gwas studies.}

\item{variantid}{String: Open Target Genetics generated id for variant (CHR_POSITION_REFALLELE_ALT_ALLELE or rsId).}
}
\value{
Returns a data frame of the colocalisation information for a lead variant in a specific study.
The table consists of the following columns:

\enumerate{
\item qtlStudyName
\item phenotypeId
\item gene.id
\item gene.symbol
\item name
\item indexVariant.id
\item indexVariant.rsId
\item beta
\item h4
\item h3
\item log2h4h3
}
}
\description{
Fetches various colocalisation  data  on molecular quantitative trait locus (QTL) analysis for a variant in a specific study.
}
\examples{
\dontrun{
otargen::qtlColocalisationVariantQuery(studyid="GCST90002357",variantid="1_154119580_C_A")
otargen::qtlColocalisationVariantQuery(studyid="GCST90002357",variantid="rs2494663")
}
}
