% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qtlColocalisationVariantQuery.R
\name{qtlColocalisationVariantQuery}
\alias{qtlColocalisationVariantQuery}
\title{Retrieves colocalisation data for a variant in a study}
\usage{
qtlColocalisationVariantQuery(studyid, variantid)
}
\arguments{
\item{\emph{studyid}}{String: Open Targets Genetics generated ID for GWAS studies.}

\item{\emph{variantid}}{String: Open Targets Genetics generated ID for variant (CHRPOSITION_REFALLELE_ALTALLELE or rsId).}
}
\value{
Returns a data frame of the colocalisation information for a lead variant in a specific study.
The table consists of the following columns:
\itemize{
  \item{\code{qtlStudyName}:} \emph{Character vector}. QTL study name.
  \item{\code{phenotypeId}:} \emph{Character vector}. Phenotype ID.
  \item{\code{gene.id}:} \emph{Character vector}. Gene ID.
  \item{\code{gene.symbol}:} \emph{Character vector}. Gene symbol.
  \item{\code{name}:} \emph{Character vector}. Tissue name.
  \item{\code{indexVariant.id}:} \emph{Character vector}. Index variant ID.
  \item{\code{indexVariant.rsId}:} \emph{Character vector}. Index variant rsID.
  \item{\code{beta}:} \emph{Numeric}. Beta value.
  \item{\code{h4}:} \emph{Numeric}. h4 value.
  \item{\code{h3}:} \emph{Numeric}. h3 value.
  \item{\code{log2h4h3}:} \emph{Numeric}. Log2(h4/h3) value.
}
}
\description{
Fetches various colocalisation data on molecular quantitative trait locus
(QTL) analysis for a variant in a specific study.
}
\examples{
\dontrun{
result <- qtlColocalisationVariantQuery(studyid = "GCST90002357", variantid = "1_154119580_C_A")
result <- qtlColocalisationVariantQuery(studyid = "GCST90002357", variantid = "rs2494663")
}
}
