% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qtlCredibleSet.R
\name{qtlCredibleSet}
\alias{qtlCredibleSet}
\title{Retrieves QTL credible set data}
\usage{
qtlCredibleSet(studyid, variantid, gene, biofeature)
}
\arguments{
\item{studyid}{String: Open Target Genetics generated id for GWAS study.}

\item{variantid}{String: Open Target Genetics generated id for variant (CHR_POSITION_REFALLELE_ALT_ALLELE or rsId).}

\item{gene}{String: a gene ENSEMBL id or a gene name.}

\item{biofeature}{String: represents either a tissue, cell type, aggregation type, protein type etc.}
}
\value{
Returns a data frame of results from qtl credible set of variants consisting of the following columns:

\enumerate{
\item tagVariant.id
\item tagVariant.rsId
\item pval
\item se
\item beta
\item postProb
\item MultisignalMethod
\item logABF
\item is95
\item is99
}
}
\description{
For a specific gene, lead variant, study id and biofeature, the function obtains tag variant information
that are considered to have a high probability of being truly associated with the trait and the corresponding scores.
}
\examples{
\dontrun{
otargen::qtlCredibleSet(studyid="Braineac2", variantid="1_55053079_C_T", gene="ENSG00000169174", biofeature="SUBSTANTIA_NIGRA")
otargen::qtlCredibleSet(studyid="Braineac2", variantid="rs7552841",gene="PCSK9", biofeature="SUBSTANTIA_NIGRA")
}

}
