% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qtlCredibleSet.R
\name{qtlCredibleSet}
\alias{qtlCredibleSet}
\title{Get qtl credible set data}
\usage{
qtlCredibleSet(studyId, variantId, geneId, bioFeature)
}
\arguments{
\item{studyid}{is the Open Target Genetics generated id for gwas studies.}

\item{variantid}{is Open Target Genetics generated id for each variant in the databse.}

\item{geneid}{is the Ensembl identifier.}

\item{biofeature}{represents either a tissue, cell type, aggregation type, protein type etc.}
}
\value{
A data frame of results from qtl credible set of variants for a given study, variant, gene and biofeature.
}
\description{
Get qtl credible set data
}
\examples{
qtlCredibleSet("Braineac2", "1_55053079_C_T", "ENSG00000169174", "SUBSTANTIA_NIGRA")
}
