% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/studiesAndLeadVariantsForGene.R
\name{studiesAndLeadVariantsForGene}
\alias{studiesAndLeadVariantsForGene}
\title{Retrieves all studies and lead variants for a gene}
\usage{
studiesAndLeadVariantsForGene(genes)
}
\arguments{
\item{genes}{String: one or more gene ENSEMBL id or gene name.}
}
\value{
Returns a dataframe
}
\description{
Retrieves all studies and lead variants for a gene
}
\examples{
\dontrun{
otargen::studiesAndLeadVariantsForGene(genes = list("ENSG00000163946","ENSG00000169174", "ENSG00000143001"))
otargen::studiesAndLeadVariantsForGene(genes = "ENSG00000169174")
otargen::studiesAndLeadVariantsForGene(genes = list("PCSK9","TASOR", "TMEM61"))
otargen::studiesAndLeadVariantsForGene(genes = "PCSK9")
}

}
