% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/studiesAndLeadVariantsForGene.R
\name{studiesAndLeadVariantsForGene}
\alias{studiesAndLeadVariantsForGene}
\title{Get all studies and lead variants for a gene}
\usage{
studiesAndLeadVariantsForGene(ensmbl_ids)
}
\arguments{
\item{ensmbl_ids}{is a identification id for genes by ensembl database}
}
\value{
a dataframe including the queried gene indentity and its colocalization data
}
\description{
Get all studies and lead variants for a gene
}
