% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/studiesAndLeadVariantsForGene.R
\name{studiesAndLeadVariantsForGene}
\alias{studiesAndLeadVariantsForGene}
\title{Get all studies and lead variants for a gene}
\usage{
studiesAndLeadVariantsForGene(ensmbl_ids)
}
\arguments{
\item{ensmbl_ids}{is a identification id for genes by ensembl database.}
}
\value{
A dataframe including the queried gene identity and its colocalization data
}
\description{
Get all studies and lead variants for a gene
}
\examples{
studiesAndLeadVariantsForGene(list("ENSG00000163946","ENSG00000169174", "ENSG00000143001"))
studiesAndLeadVariantsForGene("ENSG00000169174")
}
