% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/studiesAndLeadVariantsForGeneByL2G.R
\name{studiesAndLeadVariantsForGeneByL2G}
\alias{studiesAndLeadVariantsForGeneByL2G}
\title{Query L2G model summary data for a gene}
\usage{
studiesAndLeadVariantsForGeneByL2G(ensmbl_ids)
}
\arguments{
\item{ensmbl_ids}{is a identification id for genes by ensembl database}
}
\value{
A dataframe including the queried gene indentity and its colocalization data for L2G model
}
\description{
Query L2G model summary data for a gene
}
\examples{
studiesAndLeadVariantsForGeneByL2G(list("ENSG00000163946","ENSG00000169174", "ENSG00000143001"))
studiesAndLeadVariantsForGeneByL2G("ENSG00000169174")
}
