% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/studyInfo.R
\name{studyInfo}
\alias{studyInfo}
\title{Retrieves GWAS study id information.}
\usage{
studyInfo(studyid)
}
\arguments{
\item{\emph{studyid}}{String: Open Targets Genetics generated id for a GWAS study.}
}
\value{
Returns a tibble data table containing the GWAS study information. The columns in the table are as follows:

\enumerate{
\item studyId
\item traitReported
\item source
\item traitEfos
\item pmid
\item pubDate
\item pubJournal
\item pubTitle
\item pubAuthor
\item hasSumstats
\item ancestryInitial
\item nInitial
\item nReplication
\item traitCategory
\item numAssocLoci
\item nTotal
}
}
\description{
For a given study id, this function returns all the  relevant information about
GWAS study such as PubMed id, studied trait EFO id, case/control size etc.
}
\examples{
\dontrun{
otargen::studyInfo(studyid = "GCST90002357")
}
}
