% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/studyInfo.R
\name{studyInfo}
\alias{studyInfo}
\title{Gets the information about the input study id which links top loci with a trait.}
\usage{
studyInfo(studyid)
}
\arguments{
\item{studyid}{is the Open Target Genetics generated id for gwas studies.}
}
\value{
A dataframe containing the study information.
}
\description{
Gets the information about the input study id which links top loci with a trait.
}
\examples{
studyInfo("GCST90002357")
}
