% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/studyLocus2GeneTable.R
\name{studyLocus2GeneTable}
\alias{studyLocus2GeneTable}
\title{Get locus data table of a variant around a gene}
\usage{
studyLocus2GeneTable(studyid, variantid)
}
\arguments{
\item{studyid}{is the Open Target Genetics generated id for gwas studies.}

\item{variantid}{is Open Target Genetics generated id for each variant in the databse.}
}
\value{
a datframe with summary stats of the study and datatable of the various calculated
scores and features for any lead variant
}
\description{
Get locus data table of a variant around a gene
}
