% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/studyLocus2GeneTable.R
\name{studyLocus2GeneTable}
\alias{studyLocus2GeneTable}
\title{Retrieves L2G data table.}
\usage{
studyLocus2GeneTable(studyid, variantid)
}
\arguments{
\item{\emph{studyid}}{String: Open Targets Genetics generated ID for a GWAS study.}

\item{\emph{variantid}}{String: Open Targets Genetics generated ID for a variant (CHRPOSITION_REFALLELE_ALTALLELE or rsId).}
}
\value{
Returns a data frame with the summary statistics of the study and a data table containing various calculated scores and features for any lead variant.
The table consists of the following columns:
\itemize{
  \item{\code{studyId}:} \emph{Character}. Study ID.
  \item{\code{variant.id}:} \emph{Character}. Variant ID.
  \item{\code{variant.rsId}:} \emph{Character}. Variant rsID.
  \item{\code{yProbaDistance}:} \emph{Numeric}. Distance score.
  \item{\code{yProbaModel}:} \emph{Numeric}. Model score.
  \item{\code{yProbaMolecularQTL}:} \emph{Numeric}. Molecular QTL score.
  \item{\code{yProbaPathogenicity}:} \emph{Numeric}. Pathogenicity score.
  \item{\code{yProbaInteraction}:} \emph{Numeric}. Interaction score.
  \item{\code{hasColoc}:} \emph{Logical}. Indicates if colocalization data is available.
  \item{\code{distanceToLocus}:} \emph{Numeric}. Distance to the locus.
  \item{\code{gene.id}:} \emph{Character}. Gene ID.
  \item{\code{gene.symbol}:} \emph{Character}. Gene symbol.
}
}
\description{
Fetches the locus-to-gene (L2G) pipeline summary data table for
the neighboring genes of a variant in a GWAS study.
}
\examples{
\dontrun{
result <- studyLocus2GeneTable(studyid = "GCST90002357", variantid = "1_154119580_C_A")
result <- studyLocus2GeneTable(studyid = "GCST90002357", variantid = "rs2494663")
}
}
