% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/studyLocus2GeneTable.R
\name{studyLocus2GeneTable}
\alias{studyLocus2GeneTable}
\title{Retrieves L2G data table.}
\usage{
studyLocus2GeneTable(studyid, variantid)
}
\arguments{
\item{\emph{studyid}}{String: Open Target Genetics generated id for GWAS study.}

\item{\emph{variantid}}{String: Open Target Genetics generated id for variant (CHR_POSITION_REFALLELE_ALT_ALLELE or rsId).}
}
\value{
Returns a data frame with summary stats of the study and data table of the various calculated scores and features for any lead variant.
The table consists of the following columns:

\enumerate{
\item studyId
\item variant.id
\item variant.rsId
\item yProbaDistance
\item yProbaModel
\item yProbaMolecularQTL
\item yProbaPathogenicity
\item yProbaInteraction
\item hasColoc
\item distanceToLocus
\item gene.id
\item gene.symbol
}
}
\description{
Fetches the locus-to-gene (L2G) pipeline summary data table for
the neighboring genes of a variant in a GWAS study.
}
\examples{
\dontrun{
otargen::studyLocus2GeneTable(studyid = "GCST90002357", variantid = "1_154119580_C_A")
otargen::studyLocus2GeneTable(studyid = "GCST90002357", variantid = "rs2494663")
}
}
