% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/studyVariants.R
\name{studyVariants}
\alias{studyVariants}
\title{Retrieves variants information for a study.}
\usage{
studyVariants(studyid)
}
\arguments{
\item{studyid}{String: Open Targets Genetics generated id for GWAS study.}
}
\value{
Returns a list of data frames.

The loci data consists of the following columns:

\enumerate{
\item variant.id
\item pval
\item variant.nearestCodingGene.symbol
\item variant.rsId
\item variant.chromosome
\item variant.position
\item variant.nearestCodingGeneDistance
\item credibleSetSize
\item ldSetSize
\item oddsRatio
\item beta
}

The information about the loci genes consists of the following columns:

\enumerate{
\item score
\item gene.id
\item gene.symbol
}
}
\description{
For an input studyid, the function gives out information for all variants for all
associated loci and information about the loci genes
}
\examples{
\dontrun{
otargen::studyVariants(studyid = "GCST003155")
}
}
