% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tagVariantsAndStudiesForIndexVariant.R
\name{tagVariantsAndStudiesForIndexVariant}
\alias{tagVariantsAndStudiesForIndexVariant}
\title{Retrieves tag variants and studies for a given index variant.}
\usage{
tagVariantsAndStudiesForIndexVariant(variantid, pageindex = 0, pagesize = 20)
}
\arguments{
\item{variantid}{String: Open Target Genetics generated id for variant (CHR_POSITION_REFALLELE_ALT_ALLELE or rsId).}

\item{pageIndex}{Int: Index of the current page, pagination index >= 0.}

\item{pageSize}{Int: No. of records in a page, pagination size > 0.}
}
\value{
Data Frame containing the variant associations connected to the input index variant with the above
mentioned columns.
}
\description{
For an input index variant id, a table is generated with the following columns -
tagVariant.id, tagVariant.chromosome, tagVariant.rsId, tagVariant.position, study.studyId,
study.traitReported, study.traitCategory, pval, pvalMantissa, pvalExponent, nTotal, nCases,
overallR2, afr1000GProp, amr1000GProp, eas1000GProp, eur1000GProp, sas1000GProp, oddsRatio,
oddsRatioCILower, oddsRatioCIUpper, posteriorProbability, beta, betaCILower, betaCIUpper, direction, log10Abf.
}
\examples{

tagVariantsAndStudiesForIndexVariant(variantid = "1_109274968_G_T")
or
tagVariantsAndStudiesForIndexVariant(variantid = "1_109274968_G_T", pageindex = 1, pagesize = 50)

}
