% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tagVariantsAndStudiesForIndexVariant.R
\name{tagVariantsAndStudiesForIndexVariant}
\alias{tagVariantsAndStudiesForIndexVariant}
\title{Retrieves tag variants and studies for a given index variant.}
\usage{
tagVariantsAndStudiesForIndexVariant(variantid, pageindex = 0, pagesize = 20)
}
\arguments{
\item{\emph{variantid}}{String: Open Target Genetics generated id for variant (CHR_POSITION_REFALLELE_ALT_ALLELE or rsId).}

\item{\emph{pageindex}}{Int: Index of the current page, pagination index >= 0.}

\item{\emph{pagesize}}{Int: No. of records in a page, pagination size > 0.}
}
\value{
Returns a data Frame containing the variant associations connected to the input index variant with the following columns:

\enumerate{
\item tagVariant.id
\item tagVariant.chromosome
\item tagVariant.rsId
\item tagVariant.position
\item study.studyId
\item study.traitReported
\item study.traitCategory
\item pval
\item pvalMantissa
\item pvalExponent
\item nTotal
\item nCases
\item overallR2
\item afr1000GProp
\item amr1000GProp
\item eas1000GProp
\item eur1000GProp
\item sas1000GProp
\item oddsRatio
\item oddsRatioCILower
\item oddsRatioCIUpper
\item posteriorProbability
\item beta
\item betaCILower
\item betaCIUpper
\item direction
\item log10Abf
}
}
\description{
For an input index variant id, fetches the tag variant information
along with study information and scores.
}
\examples{
\dontrun{
otargen::tagVariantsAndStudiesForIndexVariant(variantid = "1_109274968_G_T")
otargen::tagVariantsAndStudiesForIndexVariant(variantid = "1_109274968_G_T"
    ,pageindex = 1, pagesize = 50)
}
}
