% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topOverlappedStudies.R
\name{topOverlappedStudies}
\alias{topOverlappedStudies}
\title{Gives out top studies ordered by loci overlap.}
\usage{
topOverlappedStudies(studyid, pageindex = 0, pagesize = 20)
}
\arguments{
\item{studyid}{which links the top loci with a trait. Formet: String}

\item{pageIndex}{pagination index >= 0. Index of the current page.}

\item{pageSize}{pagination size > 0. No. of records in a page. Default: 20}
}
\value{
A data frame with top studies.
}
\description{
Gives out top studies ordered by loci overlap.
}
\examples{
topOverlappedStudies("GCST006614_3")
topOverlappedStudies("NEALE2_6177_1", pageindex=1, pagesize=50)

}
